reorganizing folder structure and added create table - need to actually add to database
