+++ compareApp/client/main.html
@@ -1,10 +1,14 @@
 <head>
-  <title>CompareApp Draft</title>
+    <title>CompareApp Draft</title>
 
-<script src="http://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
-<script src="http://netdna.bootstrapcdn.com/bootstrap/3.0.0/js/bootstrap.min.js"></script>
+    <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
+    <script src="http://netdna.bootstrapcdn.com/bootstrap/3.0.0/js/bootstrap.min.js"></script>
 </head>
- 
+
 <body>
-  <div id="render-target"></div>
+    <div id="render-target">
+        <div id="header-container"></div>
+        <div id="app-container"></div>
+        <div id="footer-container"></div>
+    </div>
 </body>
