+++ compareApp/imports/ui/Menu/DataDelete.jsx
